Complete code cleanup and consolidation
Major improvements:
- Fix f32 constant precision warnings (golden ratio, sqrt(3))
- Consolidate consts_tests.rs into consts.rs for better maintainability
- Remove unused type alias F64 from ellipse.rs
- Fix Clippy warnings with compound assignment operators (+=, -=, *=)
- Delete empty metrics folder and update module references
- All 477 tests passing with clean warnings-free build

Technical details:
- Special constants: golden_ratio_f32 = 1.6180340, sqrt_3_f32 = 1.7320508
- Test consolidation: 48 lines moved from separate file to main module
- Clippy fixes: e_k -= term, power *= e_squared, sum += integrand * dt
- Module cleanup: removed metrics folder, updated lib.rs references
- Type safety: proper f64/f32 comparisons in tolerance tests
