+++ foundation/analysis/src/lib.rs
@@ -8,11 +8,11 @@ pub mod consts;
 // pub mod curves; - 楕円機能は geo_core に統合されました
 pub mod geometry; // 純粋数学的幾何図形
 pub mod linalg; // 高速線形代数モジュール（数値解析専用）
-// pub mod numerical_methods; - numerics/solver に移動済み
+                // pub mod numerical_methods; - numerics/solver に移動済み
 
 // 新しい分類構造
 pub mod approximations; // 幾何学的近似計算
-pub mod metrics; // 距離・計量計算
+                        // pub mod metrics; - 距離・計量計算は numerics に統合されました
 pub mod numerics; // 数値計算基盤
 
 // pub mod sampling; - 機能は geo_algorithms/sampling.rs に移動済み
