--- foundation/analysis/src/metrics/mod.rs
@@ -1,4 +1,0 @@
-//! 距離・計量計算モジュール
-//!
-//! ベクトル・距離計算は numerics モジュールに移動しました。
-//! 今後は crate::numerics を使用してください。
